# Conflicts:
#	config.json
#	includes/ImageUtil.h
#	src/ImageUtil.cpp
#	src/main.cpp

+++ src/Encode.cpp
@@ -80,7 +80,7 @@ double getLambdaMode(int qp){
  */
 double getLambdaPred(int qp, double k){
     extern bool lambda_inject_flag;
-    extern int injected_lambda;
+    extern double injected_lambda;
 
     if(lambda_inject_flag) return injected_lambda;
